Rename hybrid_execution to htbrid_exec
