+++ src/hybrid/hybrid_exec.h
